+++ Manbil/Rendering/GUI/GUI Elements/GUITexture.h
@@ -16,19 +16,19 @@ public:
     bool IsButton;
 
     //Only applicable if "IsButton" is true. Raised when this element is clicked.
-    void(*OnClicked)(GUITexture * clicked, Vector2i localMouse, void* pData) = 0;
+    void(*OnClicked)(GUITexture * clicked, Vector2f localMouse, void* pData) = 0;
     //Only applicable if "IsButton" is true. Raised when the mouse releases after clicking this element.
-    void(*OnReleased)(GUITexture * released, Vector2i localMouse, void* pData) = 0;
+    void(*OnReleased)(GUITexture * released, Vector2f localMouse, void* pData) = 0;
     //Only applicable if "IsButton" is true.
     void *OnClicked_pData = 0,
          *OnReleased_pData = 0;
 
 
-    virtual Vector2i GetCollisionCenter(void) const override { return center; }
-    virtual Vector2i GetCollisionDimensions(void) const override;
+    virtual Vector2f GetCollisionCenter(void) const override { return center; }
+    virtual Vector2f GetCollisionDimensions(void) const override;
 
-    virtual void MoveElement(Vector2i moveAmount) override { center += moveAmount; }
-    virtual void SetPosition(Vector2i newPos) override { center = newPos; }
+    virtual void MoveElement(Vector2f moveAmount) override { center += moveAmount; }
+    virtual void SetPosition(Vector2f newPos) override { center = newPos; }
 
     virtual void ScaleBy(Vector2f scaleAmount) override { Scale.MultiplyComponents(scaleAmount); }
     virtual void SetScale(Vector2f newScale) override { Scale = newScale; }
@@ -41,12 +41,12 @@ public:
 
     virtual std::string Render(float elapsedTime, const RenderInfo & info) override;
 
-    virtual void OnMouseClick(Vector2i mouse_centerOffset) override;
-    virtual void OnMouseRelease(Vector2i mouse_centerOffset) override;
+    virtual void OnMouseClick(Vector2f mouse_centerOffset) override;
+    virtual void OnMouseRelease(Vector2f mouse_centerOffset) override;
 
     
 private:
 
-    Vector2i center;
+    Vector2f center;
     bool isBeingClicked;
 };
